did Ex_1.2 as a multi file, single package java App. I'm proud to announce mastery of static imports: 	package test; 	import static test.Ansi.*; 	import static test.Scan.*; 	import static test.Debug.*; 	import java.util.Scanner;
the Java compiler (javac) is, however, annoying and forced me to initilise my var r (App.java:25), as it did somehow not understand the variable __was__ neceserally assigned to something before arriving at Java.app:53...
	Fuck that.
		might think about pre-initializing variables to default states, but Uck.

+++ Part_1/Ex_1.2/src/test/Scan.java
@@ -37,7 +37,7 @@ public final class Scan {
 		}
 	}
 
-	public static double scan_int(Scanner sc, String prompt)
+	public static int scan_int(Scanner sc, String prompt)
 	{
 		int r;
 
@@ -62,7 +62,7 @@ public final class Scan {
 		}
 	}
 
-	public static double scan_float(Scanner sc, String prompt)
+	public static float scan_float(Scanner sc, String prompt)
 	{
 		float r;
